+++ R/dispatcher.R
@@ -155,7 +155,8 @@ dispatcher <- function(host, url = NULL, n = NULL, ...,
     dial_and_sync_socket(sock = sockc, url = monitor, asyncdial = asyncdial)
     recv(sockc, mode = 5L, block = .timelimit) && stop(.messages[["sync_timeout"]])
     send_aio(sockc, c(Sys.getpid(), servernames), mode = 2L)
-    cmessage <- recv_aio_signal(sockc, cv = cv, mode = 5L)
+    ctrctx <- .context(sockc)
+    cmessage <- recv_aio_signal(ctrctx, cv = cv, mode = 5L)
   }
 
   suspendInterrupts(
@@ -199,8 +200,9 @@ dispatcher <- function(host, url = NULL, n = NULL, ...,
         } else {
           data <- as.integer(c(seq_n, activevec, instance, assigned, complete))
         }
-        send_aio(sockc, data = data, mode = 2L)
-        cmessage <- recv_aio_signal(sockc, cv = cv, mode = 5L)
+        ctrsaio <- send_aio(ctrctx, data = data, mode = 2L)
+        ctrctx <- .context(sockc)
+        cmessage <- recv_aio_signal(ctrctx, cv = cv, mode = 5L)
         next
       }
 
@@ -309,7 +311,7 @@ query_dispatcher <- function(sock, command, mode) {
 }
 
 query_status <- function(envir) {
-  res <- query_dispatcher(sock = envir[["sockc"]], command = 0L, mode = 5L)
+  res <- query_dispatcher(sock = .context(envir[["sockc"]]), command = 0L, mode = 5L)
   is.object(res) && return(res)
   `attributes<-`(res, list(dim = c(envir[["n"]], 5L),
                            dimnames = list(envir[["urls"]], c("i", "online", "instance", "assigned", "complete"))))